+++ src/Attribute/HttpClientConfiguration.php
@@ -8,6 +8,7 @@ final readonly class HttpClientConfiguration
     public function __construct(
         public float $timeout = 10,
         public int $retry = 0,
+        public bool $allowSelfSignedCertificate = false,
     ) {
     }
 }